+++ packages/core/src/visitors.ts
@@ -277,7 +277,7 @@ type ArazzoFlatVisitor = {
   ExtendedOperation?: VisitFunctionOrObject<ExtendedOperation>;
   Replacement?: VisitFunctionOrObject<Replacement>;
   RequestBody?: VisitFunctionOrObject<RequestBody>;
-  CriteriaObject?: VisitFunctionOrObject<CriteriaObject>;
+  CriterionObject?: VisitFunctionOrObject<CriterionObject>;
   OnSuccessObject?: VisitFunctionOrObject<OnSuccessObject>;
   OnFailureObject?: VisitFunctionOrObject<OnFailureObject>;
   Step?: VisitFunctionOrObject<Step>;