+++ packages/respect-core/src/modules/flow-runner/get-test-description-from-file.ts
@@ -47,11 +47,12 @@ export async function bundleArazzo(options: BundleArazzoOptions) {
     extends: ['recommended-strict'],
     arazzo1Rules: {
       'no-criteria-xpath': 'error',
-      'respect-supported-versions': 'warn',
-      'no-x-security-scheme-name-without-openapi': 'error',
-      'x-security-scheme-required-values': 'error',
-      'x-security-scheme-name-reference': 'error',
       'no-x-security-both-scheme-and-scheme-name': 'error',
+      'no-x-security-scheme-name-without-openapi': 'error',
+      'outputs-defined': 'error',
+      'respect-supported-versions': 'warn',
+      'x-security-scheme-name-reference': 'error',
+      'x-security-scheme-required-values': 'error',
     },
   });