+++ packages/respect-core/src/modules/flow-runner/success-criteria/validate-success-criteria.ts
@@ -1,19 +1,19 @@
 import { isPlainObject } from '@redocly/openapi-core';
 
 import {
-  type CriteriaObject,
+  type CriterionObject,
   type RegexpSuccessCriteria,
   type JsonPathSuccessCriteria,
 } from '../../../types.js';
 
 export function isRegexpSuccessCriteria(
-  criteria: CriteriaObject
+  criteria: CriterionObject
 ): criteria is RegexpSuccessCriteria {
   return (criteria as RegexpSuccessCriteria).type === 'regex';
 }
 
 export function isJSONPathSuccessCriteria(
-  criteria: CriteriaObject
+  criteria: CriterionObject
 ): criteria is JsonPathSuccessCriteria {
   const typeValue = (criteria as JsonPathSuccessCriteria)?.type;
 
@@ -34,8 +34,8 @@ export const ALLOWED_EXPRESSION_KEYS = [
   '$components',
 ];
 
-export function validateSuccessCriteria(successCriteria: CriteriaObject[]) {
-  successCriteria.forEach((criteria: CriteriaObject) => {
+export function validateSuccessCriteria(successCriteria: CriterionObject[]) {
+  successCriteria.forEach((criteria: CriterionObject) => {
     const { condition } = criteria;
 
     if (isRegexpSuccessCriteria(criteria)) {